move file to examples
